:sweep: move org config for clarity
