+++ src/services/orgConfig.ts
@@ -1,0 +1,13 @@
+export type GitHubTeamName = string;
+
+export type ManagedGitHubTeam = {
+    Name: GitHubTeamName,
+    DisplayName?: string
+}
+
+export type OrgConfiguration = {
+    GitHubTeamNames?: string[]
+    Teams?: ManagedGitHubTeam[]
+    OrganizationMembersGroup?: string
+    OrganizationOwnersGroup?: string | ManagedGitHubTeam
+}
